allow cache skipping, going to test on monolith

+++ smallab/experiment_types/handlers/checkpointed_experiment_handler.py
@@ -8,7 +8,7 @@ from dateutil.parser import parse
 
 from smallab.dashboard.dashboard_events import ProgressEvent
 from smallab.dashboard.utils import put_in_event_queue
-from smallab.experiment_types.checkpointed_experiment import CheckpointedExperiment
+from smallab.experiment_types.checkpointed_experiment import CheckpointedExperiment, HasCheckpoint
 from smallab.experiment_types.handlers.base_handler import BaseHandler
 from smallab.file_locations import get_partial_save_directory
 from smallab.smallab_types import Specification
@@ -90,21 +90,25 @@ class CheckpointedExperimentHandler(BaseHandler):
         return checkpoints
 
     def _save_checkpoint(self, experiment, name, specification):
-        experiment_hash = specification_hash(specification)
-        checkpoint_name = str(datetime.datetime.now())
-        try:
-            location = get_partial_save_directory(name, specification)
-            os.makedirs(location, exist_ok=True)
-            # TODO make sure a checkpoint with this name doesn't already exist
-            with open(os.path.join(location, checkpoint_name + ".pkl"), "wb") as f:
-                dill.dump(experiment, f)
-            logging.getLogger("smallab.{id}.checkpoint".format(id=experiment_hash)).info(
-                "Succesfully checkpointed {chp}".format(chp=checkpoint_name))
-            checkpoints = os.listdir(location)
-            if len(checkpoints) > self.rolled_backups:
-                checkpoints = self._get_time_sorted_checkpoints(name, specification)
-                os.remove(os.path.join(location, str(checkpoints[0]) + ".pkl"))
-        except:
-            logging.getLogger("smallab.{id}.checkpoint".format(id=experiment_hash)).warning(
-                "Unsuccesful checkpoint {chp}".format(chp=checkpoint_name),
-                exc_info=True)
+        assert isinstance(experiment,HasCheckpoint)
+        experiment.set_steps_since_checkpoint(experiment.get_steps_since_checkpiont() + 1)
+        if experiment.get_steps_since_checkpiont() >= experiment.steps_before_checkpoint():
+            experiment.set_steps_since_checkpoint(0)
+            experiment_hash = specification_hash(specification)
+            checkpoint_name = str(datetime.datetime.now())
+            try:
+                location = get_partial_save_directory(name, specification)
+                os.makedirs(location, exist_ok=True)
+                # TODO make sure a checkpoint with this name doesn't already exist
+                with open(os.path.join(location, checkpoint_name + ".pkl"), "wb") as f:
+                    dill.dump(experiment, f)
+                logging.getLogger("smallab.{id}.checkpoint".format(id=experiment_hash)).info(
+                    "Succesfully checkpointed {chp}".format(chp=checkpoint_name))
+                checkpoints = os.listdir(location)
+                if len(checkpoints) > self.rolled_backups:
+                    checkpoints = self._get_time_sorted_checkpoints(name, specification)
+                    os.remove(os.path.join(location, str(checkpoints[0]) + ".pkl"))
+            except:
+                logging.getLogger("smallab.{id}.checkpoint".format(id=experiment_hash)).warning(
+                    "Unsuccesful checkpoint {chp}".format(chp=checkpoint_name),
+                    exc_info=True)
